requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
psycopg2-binary>=2.9.9
Pillow>=10.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7